requirements.txt
fastapi
uvicorn
pandas
beautifulsoup4
tiktoken
groq
tqdm
python-dotenv
python-multipart